+++ CHANGELOG.md
@@ -37,6 +37,7 @@ v# Changelog
 ### Changed
 * `write_err` (`output`) has been improved to be more consistent in the output between the different kind of errors.  The names of the arguments have been changed to be (hopefully) clearer.
 * Module `geometry` now uses the new `runstat` from `exception` to manage exceptions.
+* Improved error messages in `mcd_tensor`.
 
 ### Improved
 * Better efficiency of `int_xn_e2ax2` (`math`).